requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.16.2
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
asyncpg==0.30.0
    # via -r requirements.in
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
click==8.2.1
    # via uvicorn
coverage[toml]==7.9.1
    # via pytest-cov
fastapi==0.115.13
    # via -r requirements.in
greenlet==3.2.3
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
iniconfig==2.1.0
    # via pytest
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via mako
packaging==25.0
    # via pytest
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
prometheus-client==0.22.1
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via -r requirements.in
pygments==2.19.1
    # via pytest
pytest==8.4.1
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-cov
pytest-asyncio==1.0.0
    # via -r requirements.in
pytest-cov==6.2.1
    # via -r requirements.in
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via -r requirements.in
pyyaml==6.0.2
    # via uvicorn
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.41
    # via
    #   -r requirements.in
    #   alembic
starlette==0.46.2
    # via fastapi
structlog==25.4.0
    # via -r requirements.in
typing-extensions==4.14.0
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
uvicorn[standard]==0.34.3
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
